test(authorization): extract shared setup and apply Arrange/Act/Assert
Move the duplicated app-building, mock-wiring, and route-registration
ceremony out of authorization.test.ts and permissions.startup-guard.test.ts
into src/test-support/ helpers, and restructure each test into a single
Arrange/Act/Assert flow (splitting the one two-act PATCH test in two).
